+++ server/internal/models/user_games.go
@@ -10,9 +10,9 @@ const (
 )
 
 type UserGames struct {
-	ID       int64      `json:"id" gorm:"primary_key"`
-	UserID   int64      `json:"user_id"`
-	GameID   int64      `json:"game_id"`
+	ID       int        `json:"id" gorm:"primary_key"`
+	UserID   int        `json:"user_id"`
+	GameID   int        `json:"game_id"`
 	Priority int        `json:"priority"`
 	Status   GameStatus `json:"status" gorm:"type:varchar(20);default:'planned'"`
 }
